ai_module/pyproject.toml
[project]
name = "ai-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Valdir Junior", email = "valdir.junior.dev@gmail.com" }
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi==0.135.3",
    "google-generativeai==0.8.6",
    "httpx==0.28.1",
    "openai==2.30.0",
    "pillow==12.2.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pymupdf==1.27.2.2",
    "python-json-logger==4.1.0",
    "python-multipart==0.0.24",
    "uvicorn==0.44.0",
]
[dependency-groups]
dev = [
    "mypy==1.20.0",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "ruff==0.15.9",
]
[project.scripts]
ai-modulo = "ai_module.main:main"
ai-modulo-dev = "ai_module.main:dev"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["ai_module"]
branch = true

[tool.coverage.report]
skip_empty = true
show_missing = true




